Глобальные изменения 1. Из-за проблем парсером для типа double изменил тип тарифа double -> int 2. Исправил логику вычисления тарифа 3. Рефакторинг кода 4. Изменил дизайн истории

+++ Tariff/Tariff/Person.cs
@@ -10,9 +10,9 @@ namespace Tariff
         public string name { get; set; }
         public string email { get; set; }
         public string password { get; set; }
-        public double hotWater { get; set; }
-        public double coldWater { get; set; }
-        public double gas { get; set; }
-        public double electricity { get; set; }
+        public int hotWater { get; set; }
+        public int coldWater { get; set; }
+        public int gas { get; set; }
+        public int electricity { get; set; }
     }
 }
